readyup|refactor: Extract JSON output mode into its own module
Moves `runMultiKitJsonMode` out of `runCommand.ts` into a `runJsonMode.ts` of its own, renamed `runJsonMode`, and takes the `toJsonOriginField` helper with it.

The runner's inline settings parameter type becomes a named `JsonRunSettings`. `runCommand` keeps the `detail ?? 'full'` normalization, so the mode receives a resolved `JsonDetail`.
